Add ability for using symlinks (#18)
This adds the option to create symlinks. Plan in future to just make it
a runtime config option instead of compile time so that it can be
enabled on windows if the user has enabled developer mode

+++ src/choosenimpkg/switcher.nim
@@ -5,6 +5,8 @@ from nimblepkg/packageinfo import getNameVersion
 
 import cliparams, common
 
+const useSymLinks {.booldefine: "useSymLinks"}: bool = false
+
 when defined(windows):
   import env
 
@@ -29,14 +31,18 @@ proc compileProxyexe() =
   let (output, exitCode) = gorgeEx(cmd)
   doAssert exitCode == 0, $(output, cmd)
 
-when not defined(noBuildProxy):
+when not (defined(noBuildProxy) or useSymLinks):
   static: compileProxyexe()
 
-const
-  proxyExe = staticRead("proxyexe".addFileExt(ExeExt))
+when not useSymLinks:
+  const proxyExe = staticRead("proxyexe".addFileExt(ExeExt))
 
 proc getInstallationDir*(params: CliParams, version: Version): string =
-  return params.getInstallDir() / ("nim-$1" % $version)
+  # Nimble does this fun little thing where it quotes the binary before
+  # running it. Since # is considered not safe for posix, the whole thing gets
+  # escaped which causes the command to fail
+  let versionStr = replace($version, "#", "hash")
+  return params.getInstallDir() / ("nim-" & versionStr)
 
 proc isVersionInstalled*(params: CliParams, version: Version): bool =
   return fileExists(params.getInstallationDir(version) / "bin" /
@@ -54,13 +60,14 @@ proc areProxiesInstalled(params: CliParams, proxies: openarray[string]): bool =
   for proxy in proxies:
     # Verify that proxy exists.
     let path = params.getProxyPath(proxy)
-    if not fileExists(path):
+    if not (fileExists(path) or (symlinkExists(path) and useSymLinks)):
       return false
 
-    # Verify that proxy binary is up-to-date.
-    let contents = readFile(path)
-    if contents != proxyExe:
-      return false
+    when not useSymLinks:
+      # Verify that proxy binary is up-to-date.
+      let contents = readFile(path)
+      if contents != proxyExe:
+        return false
 
 proc isDefaultCCInPath*(params: CliParams): bool =
   # Fixes issue #104
@@ -130,28 +137,37 @@ proc writeProxy(bin: string, params: CliParams) =
         removeDir(dir)
         display("Removed", dir, priority = HighPriority)
 
-  if symlinkExists(proxyPath):
-    let msg = "Symlink for '$1' detected in '$2'. Can I remove it?" %
-              [bin, proxyPath.splitFile().dir]
-    if not prompt(dontForcePrompt, msg): return
-    let symlinkPath = expandSymlink(proxyPath)
-    removeFile(proxyPath)
-    display("Removed", "symlink pointing to $1" % symlinkPath,
-            priority = HighPriority)
 
-  # Don't write the file again if it already exists.
-  if fileExists(proxyPath) and readFile(proxyPath) == proxyExe: return
+  if symlinkExists(proxyPath) or (useSymLinks and fileExists(proxyPath)):
+    when useSymLinks:
+      removeFile(proxyPath)
+    else:
+      let msg = "Symlink for '$1' detected in '$2'. Can I remove it?" %
+                [bin, proxyPath.splitFile().dir]
+      if not prompt(dontForcePrompt, msg): return
+      let symlinkPath = expandSymlink(proxyPath)
+      removeFile(proxyPath)
+      display("Removed", "symlink pointing to $1" % symlinkPath,
+              priority = HighPriority)
 
-  try:
-    writeFile(proxyPath, proxyExe)
-  except IOError:
-    display("Warning:", "component '$1' possibly in use, write failed" % bin, Warning,
-            priority = HighPriority)
-    return
+  let targetPath = params.getExePath(bin.addFileExt(ExeExt))
+  # Use old proxy on windows, but stick with symlinks on linux systems
+  # TODO: Test if developer mode is enabled and we can create symlinks
+  when useSymLinks:
+    createSymlink(targetPath, proxyPath)
+  else:
+    # Don't write the file again if it already exists.
+    if fileExists(proxyPath) and readFile(proxyPath) == proxyExe: return
+    try:
+      writeFile(proxyPath, proxyExe)
+    except IOError:
+      display("Warning:", "component '$1' possibly in use, write failed" % bin, Warning,
+              priority = HighPriority)
+      return
 
-  # Make sure the exe has +x flag.
-  setFilePermissions(proxyPath,
-                     getFilePermissions(proxyPath) + {fpUserExec})
+  when not useSymLinks:
+    # Make sure the exe has +x flag.
+    setFilePermissions(proxyPath, getFilePermissions(proxyPath) + {fpUserExec})
   display("Installed", "component '$1'" % bin, priority = HighPriority)
 
   # Check whether this is in the user's PATH.
@@ -166,7 +182,7 @@ proc writeProxy(bin: string, params: CliParams) =
         "Binary '$1' isn't in your PATH. Ensure that '$2' is in your PATH." %
           [bin, params.getBinDir()]
     display("Hint:", msg, Warning, HighPriority)
-  elif fromPATH != "" and fromPATH != proxyPath:
+  elif fromPATH notin ["", proxyPath, targetPath]:
     display("Warning:", "Binary '$1' is shadowed by '$2'." %
             [bin, fromPATH], Warning, HighPriority)
     display("Hint:", "Ensure that '$1' is before '$2' in the PATH env var." %
